Baja física de usuario

Closes #33

See merge request proyecto2024/proyectos/grupo28/code!22

+++ admin/src/core/user/__init__.py
@@ -19,9 +19,14 @@ def create_user(**kwargs):
     return user
 
 def delete_user(user_id):
+    """
+    Baja fisica de un usuario del sistema, consultar si no sería mejor una lógica
+    """
     user = User.query.get(user_id)
     if not user:
         raise ValueError("El usuario no existe")
+    if user.system_admin:
+        raise ValueError("No puedes eliminar a un usuario administrador del sistema.")
     db.session.delete(user)
     db.session.commit()
